Item_1713
Added section to display information about an Availability Group. 1. Is part of an availability group 2. Is the connection to server or node 3. list of all nodes in cluster 4. list all the listeners in node. 5. warn of any of the nodes are off line 6. info regarding number of votes and who holds 7. is cluster using file share, disk share in quorum